Add contact page and update vite config, layouts, and 404 page
The commit includes the addition of a new contact page and updates to layouts for better website navigation. The 404 page text content has been modified and the config for Vite has been changed to allow additional files. Furthermore, website images are now being imported directly within the layouts, and a new Modal functionality has been included from SkeletonLabs for potential displaying of confirmations or notifications to the user.